create dedicated test file for directory page, move related tests over
